+++ src/ports/esp32/boards/M5STACK_CORE2/modules/m5Stack.py
@@ -1,0 +1,8 @@
+# MicroPython Helper Library for M5Stack Core2
+#
+# MIT License, Copyright (c) 2024 Sebastian Alexander Eckhardt
+#
+
+class Core2:
+    def __init__(self):
+        pass